requirements.txt
pandas==2.3.3
python-binance==1.0.32
python-dotenv==1.2.1
